==> scripts/01-format-data.Rmd <==
---
title: "Format data for occupancy modelling"
author: "Julie K. Sheard"
date: "16/04/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This is the script to format the seven data sets for occupancy modelling. Note that we use a categorical version of list length, meaning that we format the list lengths of each datasets so that data sets from The Natural History Museum of Denmark, The Natural History Museum of Aarhus, Holger Holgersen, Sämi Schär and Christian Skøtt are given list length 1, EuroAnts is given list length 2 and the Ant Hunt is given list length 4. 
                 
#Load packages
```{r}
library(devtools)
library(sparta)
library(magrittr)
library(dplyr)
library(DescTools)
library(knitr)
```

#Load and subset data
```{r}
Dataset <- read.csv("../data/Data set.csv", sep=",", header = TRUE)

NHMDK <- subset(Dataset, origin == "NHM Denmark") #8604 records
Aarhus <- subset(Dataset, origin == "NHM Aarhus") # 1163 records
Holgersen <- subset(Dataset, origin == "H. Holgersen") # 49 records
Skoett <- subset(Dataset, origin == "C. Skøtt") # 944 records
Schar <- subset(Dataset, origin == "S. Schär") # 98 records
EuroAnts <- subset(Dataset, origin == "EuroAnts") # 100 records
AntHunt <- subset(Dataset, origin == "The Ant Hunt") # 667 records
```

#Make date a date format of YYYY-MM-DD
```{r}
NHMDK$date <- as.Date(NHMDK$date)
str(NHMDK)

Aarhus$date <- as.Date(Aarhus$date)
str(Aarhus)

Holgersen$date <- as.Date(Holgersen$date)
str(Holgersen)

Skoett$date <- as.Date(Skoett$date)
str(Skoett)

Schar$date <- as.Date(Schar$date)
str(Schar)

EuroAnts$date <- as.Date(EuroAnts$date)
str(EuroAnts)

AntHunt$date <- as.Date(AntHunt$date)
str(AntHunt)
```

# Format all data sets, get rid of sites only in 1 time period
```{r}
formattedOccData_AH <- sparta::formatOccData(taxa = AntHunt$taxa, 
                                             site = AntHunt$site_10, 
                                             survey = AntHunt$date,
                                             replicate = AntHunt$GPS,
                                             closure_period = AntHunt$tp_10)
# 1 out of 667 observations will be removed as duplicates

#Summarise Ant Hunt dataset
visitsSummary(formattedOccData_AH)

formattedOccData_EU <- sparta::formatOccData(taxa = EuroAnts$taxa, 
                                             site = EuroAnts$site_10, 
                                             survey = EuroAnts$date, 
                                             replicate = EuroAnts$GPS,
                                             closure_period = EuroAnts$tp_10)
# 19 out of 43 observations will be removed as duplicates

#Summarise EuroAnts dataset
visitsSummary(formattedOccData_EU)

formattedOccData_Holgersen <- sparta::formatOccData(taxa = Holgersen$taxa, 
                                              site = Holgersen$site_10, 
                                              survey = Holgersen$date,
                                              replicate = Holgersen$GPS,
                                              closure_period = Holgersen$tp_10)
#31 out of 49 observations will be removed as duplicates

#Summarise Holgersen dataset
visitsSummary(formattedOccData_Holgersen)

formattedOccData_Schar <- sparta::formatOccData(taxa = Schar$taxa, 
                                                site = Schar$site_10, 
                                                survey = Schar$date,
                                                replicate = Schar$GPS,
                                                closure_period = Schar$tp_10)
#10 out of 98 observations will be removed as duplicates

#Summarise Schar dataset
visitsSummary(formattedOccData_Schar)


formattedOccData_Skoett <- sparta::formatOccData(taxa = Skoett$taxa, 
                                                 site = Skoett$site_10, 
                                                 survey = Skoett$date,
                                                 replicate = Skoett$GPS,
                                                 closure_period = Skoett$tp_10)
#4 out of 944 observations will be removed as duplicates

#Summarise Skoett dataset
visitsSummary(formattedOccData_Skoett)

formattedOccData_NHMDK <- sparta::formatOccData(taxa = NHMDK$taxa, 
                                              site = NHMDK$site_10, 
                                              survey = NHMDK$date,
                                              replicate = NHMDK$GPS,
                                              closure_period = NHMDK$tp_10)
#6225 out of 8604 observations will be removed as duplicates

#Summarise SNM dataset
visitsSummary(formattedOccData_NHMDK)

formattedOccData_Aarhus <- sparta::formatOccData(taxa = Aarhus$taxa, 
                                                 site = Aarhus$site_10, 
                                                 survey = Aarhus$date,
                                                 replicate = Aarhus$GPS, 
                                                 closure_period = Aarhus$tp_10)
#755 out of 1163 observations will be removed as duplicates

#Summarise Aarhus dataset
visitsSummary(formattedOccData_Aarhus)
```

#Calculate mean and median list length for each dataset
```{r}
mean(formattedOccData_Aarhus$occDetdata$L)
median(formattedOccData_Aarhus$occDetdata$L)

mean(formattedOccData_NHMDK$occDetdata$L)
median(formattedOccData_NHMDK$occDetdata$L)

mean(formattedOccData_Skoett$occDetdata$L)
median(formattedOccData_Skoett$occDetdata$L)

mean(formattedOccData_Holgersen$occDetdata$L)
median(formattedOccData_Holgersen$occDetdata$L)

mean(formattedOccData_Schar$occDetdata$L)
median(formattedOccData_Schar$occDetdata$L)

mean(formattedOccData_EU$occDetdata$L)
median(formattedOccData_EU$occDetdata$L)

mean(formattedOccData_AH$occDetdata$L)
median(formattedOccData_AH$occDetdata$L)
```

#Change time periods
SPARTA changes the tp sequentially from 1. This means all datasets except NHMDK now have wrong time periods. Change to:
Holgersen: 9
EuroAnts: 12
Ant Hunt: 12
Schar: 12
Skoett: 7, 8
Aarhus: 1, 2, 4, 5, 6, 7, 8

```{r}
formattedOccData_Schar$occDetdata$TP[formattedOccData_Schar$occDetdata$TP==1] <- 12
formattedOccData_Holgersen$occDetdata$TP[formattedOccData_Holgersen$occDetdata$TP==1] <- 9
formattedOccData_EU$occDetdata$TP[formattedOccData_EU$occDetdata$TP==1] <- 12
formattedOccData_AH$occDetdata$TP[formattedOccData_AH$occDetdata$TP==1] <- 12
formattedOccData_Skoett$occDetdata$TP[formattedOccData_Skoett$occDetdata$TP==2] <- 8
formattedOccData_Skoett$occDetdata$TP[formattedOccData_Skoett$occDetdata$TP==1] <- 7
formattedOccData_Aarhus$occDetdata$TP[formattedOccData_Aarhus$occDetdata$TP==7] <- 8
formattedOccData_Aarhus$occDetdata$TP[formattedOccData_Aarhus$occDetdata$TP==6] <- 7
formattedOccData_Aarhus$occDetdata$TP[formattedOccData_Aarhus$occDetdata$TP==5] <- 6
formattedOccData_Aarhus$occDetdata$TP[formattedOccData_Aarhus$occDetdata$TP==4] <- 5
formattedOccData_Aarhus$occDetdata$TP[formattedOccData_Aarhus$occDetdata$TP==3] <- 4
```

#Change list lengths to use categorical list lengths 1, 2 and 4. 
```{r}
#Change to 1
formattedOccData_SNM$occDetdata$L[formattedOccData_SNM$occDetdata$L<12] <- 1
formattedOccData_Aarhus$occDetdata$L[formattedOccData_Aarhus$occDetdata$L<8] <- 1
formattedOccData_Holgersen$occDetdata$L[formattedOccData_Holgersen$occDetdata$L<6] <- 1
formattedOccData_Schar$occDetdata$L[formattedOccData_Schar$occDetdata$L<7] <- 1
formattedOccData_Skoett$occDetdata$L[formattedOccData_Skoett$occDetdata$L<6] <- 1

#Change to 2
formattedOccData_EU$occDetdata$L[formattedOccData_EU$occDetdata$L<20] <- 2

#Change to 4
formattedOccData_AH$occDetdata$L[formattedOccData_AH$occDetdata$L<5] <- 3
```

```{r}
#Export individual spp_vis and occDetdata files
write.csv(formattedOccData_AH$spp_vis, '../data/spp_vis_AH.csv', row.names=FALSE)
write.csv(formattedOccData_AH$occDetdata, '../data/occDetdata_AH.csv', row.names=FALSE)

write.csv(formattedOccData_EU$occDetdata, '../data/occDetdata_EU.csv', row.names=FALSE)
write.csv(formattedOccData_EU$spp_vis, '../data/spp_vis_EU.csv', row.names=FALSE)

write.csv(formattedOccData_SNM$spp_vis, '../data/spp_vis_NHMDK.csv', row.names=FALSE)
write.csv(formattedOccData_SNM$occDetdata, '../data/occDetdata_NHMDK.csv', row.names=FALSE)

write.csv(formattedOccData_Schar$spp_vis, '../data/spp_vis_Schar.csv', row.names=FALSE)
write.csv(formattedOccData_Schar$occDetdata, '../data/occDetdata_Schar.csv', row.names=FALSE)

write.csv(formattedOccData_Skoett$spp_vis, '../data/spp_vis_Skoett.csv', row.names=FALSE)
write.csv(formattedOccData_Skoett$occDetdata, '../data/occDetdata_Skoett.csv', row.names=FALSE)

write.csv(formattedOccData_Aarhus$spp_vis, '../data/spp_vis_Aarhus.csv', row.names=FALSE)
write.csv(formattedOccData_Aarhus$occDetdata, '../data/occDetdata_Aarhus.csv', row.names=FALSE)

write.csv(formattedOccData_Holgersen$spp_vis, '../data/edited/spp_vis_Holgersen.csv', row.names=FALSE)
write.csv(formattedOccData_Holgersen$occDetdata, '../data/occDetdata_Holgersen.csv', row.names=FALSE)
```

#Merge datasets
```{r}
occDetdata_merged <- rbind(formattedOccData_AH$occDetdata, 
                           formattedOccData_NHMDK$occDetdata, 
                           formattedOccData_Aarhus$occDetdata, 
                           formattedOccData_EU$occDetdata, 
                           formattedOccData_Holgersen$occDetdata, 
                           formattedOccData_Schar$occDetdata, 
                           formattedOccData_Skoett$occDetdata)

spp_vis_merged <- dplyr::bind_rows(formattedOccData_AH$spp_vis, 
                                   formattedOccData_NHMDK$spp_vis,
                                   formattedOccData_Aarhus$spp_vis, 
                                   formattedOccData_EU$spp_vis, 
                                   formattedOccData_Holgersen$spp_vis, 
                                   formattedOccData_Schar$spp_vis, 
                                   formattedOccData_Skoett$spp_vis)

spp_vis_merged[is.na(spp_vis_merged)] <- FALSE

#Export merged occDetdata
write.csv(occDetdata_merged, '../data/occDetdata_merged.csv', row.names=FALSE)

#Export  spp_vis datasets
write.csv(spp_vis_merged, '../data/spp_vis_merged.csv', row.names=FALSE)
```
