==> scripts/02-run-model.Rmd <==
---
title: "Run full occupancy model"
author: "Julie K. Sheard"
date: "16/04/2021"
output: html_document

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This is a script to model the change in occupancy for 29 Danish ant species from 1900-2019 based on a multi-sourced data set. 
Model parameters: 
Survey: date (yyyy/mm/dd)
Closure period: 10 years
Grid cell: 10x10 km

# Initial set-up

## Load packages
```{r}
library(sparta)
library(snowfall) # Use the occupancy modelling function to parrellise the process using snowfall
```

## Load data, make sure date is a date and check for correct format
```{r}
occDetdata_merged <- read.table('../data/occDetdata_merged.csv', sep = ",", header = TRUE)

spp_vis_merged <- read.csv('../data/spp_vis_merged.csv', sep=",", header = TRUE)

head(occDetdata_merged)
head(spp_vis_merged)
```

# Run occupancy models
The model was initially run at 50,000 iterations and then rerun for those species which had not converged, doubling the number of iterations with each rerun. 

## Run first model, 50,000 iterations
```{r}
# Check computer CPUS and adjust before initialising the cluster
sfInit(parallel = TRUE, cpus = 8)

# Export my data to the cluster
sfExport('occDetdata_merged')
sfExport('spp_vis_merged')


occ_mod_function <- function(taxa_name){
  
library(sparta)
  
occ_out <- occDetFunc(taxa_name = taxa_name,
                        n_iterations = 50000, #600.000 max, 100.000 min
                        burnin = 25000, #Half of n_iterations
                        thinning = 3,
                        nyr = 2,
                        n_chains = 3, 
                        occDetdata = occDetdata_merged,
                        spp_vis = spp_vis_merged,
                        modeltype = c('ranwalk', 'halfcauchy', 
                                      'catlistlength'),
                        write_results = FALSE,
                        seed = 123)}

# Run parallel analyses of all species
system.time({
para_out_50 <- sfClusterApplyLB(c('Camponotus_herculeanus', 'Formica_cinerea', 'Formica_cunicularia',
                               'Formica_exsecta','Formica_fusca', 'Formica_picea', 
                               'Formica_polyctena','Formica_pratensis', 'Formica_pressilabris', 
                               'Formica_rufa', 'Formica_rufibarbis', 'Formica_sanguinea', 
                               'Formica_truncorum', 'Formica_uralensis', 'Formicoxenus_nitidulus', 
                               'Lasius_flavus', 'Lasius_fuliginosus', 'Lasius_meridionalis', 
                               'Lasius_niger', 'Lasius_umbratus', 'Leptothorax_acervorum', 
                               'Myrmica_lobicornis', 'Myrmica_rubra', 'Myrmica_ruginodis', 
                               'Myrmica_rugulosa', 'Myrmica_sabuleti', 'Myrmica_scabrinodis', 
                               'Myrmica_schencki', 'Tetramorium_caespitum'), 
                                occ_mod_function)
})

# Stop the cluster
sfStop()

#Save results
save(para_out_50, file="C:/Users/julie/Documents/Projects/Occupancy/output/para_out_50.RData")
```

## Rerun for unconverged species at 100,000 iterations
```{r}
# Check computer CPUS and adjust before initialising the cluster
sfInit(parallel = TRUE, cpus = 8)

# Export my data to the cluster
sfExport('occDetdata_merged')
sfExport('spp_vis_merged')


occ_mod_function <- function(taxa_name){
  
library(sparta)
  
occ_out <- occDetFunc(taxa_name = taxa_name,
                        n_iterations = 100000, #600.000 max, 100.000 min
                        burnin = 50000, #Half of n_iterations
                        thinning = 3,
                        nyr = 2,
                        n_chains = 3, 
                        occDetdata = occDetdata_merged,
                        spp_vis = spp_vis_merged,
                        modeltype = c('ranwalk', 'halfcauchy', 
                                      'catlistlength'),
                        write_results = FALSE,
                        seed = 123)}

# Run parallel analyses of all species
system.time({
para_out_100 <- sfClusterApplyLB(c('Formica_exsecta', 'Formica_fusca', 'Formica_picea', 
                               'Formica_pratensis', 'Lasius_flavus', 'Lasius_meridionalis', 
                               'Lasius_niger', 'Leptothorax_acervorum', 'Myrmica_lobicornis',
                               'Tetramorium_caespitum'), 
                                occ_mod_function)
})

# Stop the cluster
sfStop()

# Save results from cluster
save(para_out_100, file="C:/Users/julie/Documents/Projects/Occupancy/output/para_out_100.RData")
```

## Rerun for unconverged species at 200,000 iterations
```{r}
# Use the occupancy modelling function to parrellise the process using snowfall
library(snowfall)

# Check computer CPUS and adjust before initialising the cluster
sfInit(parallel = TRUE, cpus = 8)

# Export my data to the cluster
sfExport('occDetdata_merged')
sfExport('spp_vis_merged')


occ_mod_function <- function(taxa_name){
  
library(sparta)
  
occ_out <- occDetFunc(taxa_name = taxa_name,
                        n_iterations = 200000, #600.000 max, 100.000 min
                        burnin = 100000, #Half of n_iterations
                        thinning = 3,
                        nyr = 2,
                        n_chains = 3, 
                        occDetdata = occDetdata_merged,
                        spp_vis = spp_vis_merged,
                        modeltype = c('ranwalk', 'halfcauchy', 
                                      'catlistlength'),
                        write_results = FALSE,
                        seed = 123)}

# Run parallel analyses of all species
system.time({
para_out_200 <- sfClusterApplyLB(c('Formica_exsecta', 'Formica_picea', 'Lasius_meridionalis', 
                                   'Lasius_niger', 'Myrmica_lobicornis'), 
                                occ_mod_function)
})

# Stop the cluster
sfStop()

# Save results from cluster
save(para_out_200, file="C:/Users/julie/Documents/Projects/Occupancy/output/para_out_200.RData")
```

#Rerun for unconverged species at 400,000 iterations
```{r}
# Use the occupancy modelling function to parrellise the process using snowfall
library(snowfall)

# Check computer CPUS and adjust before initialising the cluster
sfInit(parallel = TRUE, cpus = 8)

# Export my data to the cluster
sfExport('occDetdata_merged')
sfExport('spp_vis_merged')


occ_mod_function <- function(taxa_name){
  
library(sparta)
  
occ_out <- occDetFunc(taxa_name = taxa_name,
                        n_iterations = 400000, #600.000 max, 100.000 min
                        burnin = 200000, #Half of n_iterations
                        thinning = 3,
                        nyr = 2,
                        n_chains = 3, 
                        occDetdata = occDetdata_merged,
                        spp_vis = spp_vis_merged,
                        modeltype = c('ranwalk', 'halfcauchy', 
                                      'catlistlength'),
                        write_results = FALSE,
                        seed = 123)}

# Run parallel analyses of all species
system.time({
para_out_400 <- sfClusterApplyLB(c('Lasius_meridionalis', 'Lasius_niger'), 
                                occ_mod_function)
})

# Stop the cluster
sfStop()

# Save results from cluster
save(para_out_400, file="C:/Users/julie/Documents/Projects/Occupancy/output/para_out_400.RData")
```


#Rerun for unconverged species at 600,000 iterations
```{r}
# Use the occupancy modelling function to parrellise the process using snowfall
library(snowfall)

# Check computer CPUS and adjust before initialising the cluster
sfInit(parallel = TRUE, cpus = 8)

# Export my data to the cluster
sfExport('occDetdata_merged')
sfExport('spp_vis_merged')


occ_mod_function <- function(taxa_name){
  
library(sparta)
  
occ_out <- occDetFunc(taxa_name = taxa_name,
                        n_iterations = 600000, #600.000 max, 100.000 min
                        burnin = 300000, #Half of n_iterations
                        thinning = 3,
                        nyr = 2,
                        n_chains = 3, 
                        occDetdata = occDetdata_merged,
                        spp_vis = spp_vis_merged,
                        modeltype = c('ranwalk', 'halfcauchy', 
                                      'catlistlength'),
                        write_results = FALSE,
                        seed = 123)}

# Run parallel analyses of all species
system.time({
para_out_600 <- sfClusterApplyLB(c('Lasius_meridionalis'), 
                                occ_mod_function)
})

# Stop the cluster
sfStop()

# Save results from cluster
save(para_out_600, file="C:/Users/julie/Documents/Projects/Occupancy/output/para_out_600.RData")
```
